:bug: Change triggers that re-render image, always check if job used current data
Previously, `render_image` was only called when:
- the image slice index changed (this is also triggered when a new data set is loaded, even if the slide index remains the same)
- the overlay toggle is used
Since this was not dependent on changes of results (e.g. a inference or training job is selected), it can happen that outdated result information is shown if the overlay toggle was not toggled in between switching jobs. Inference results from other data sets, also
Partially adresses #188, adresses #190
The drop-downs being used both as result selection and as state trackers for jobs complicates this.